WIP implementation of geophysical displacements
1. Defined a base class for geophysical displacements [WIP]
2. Moved Constants class to `constants` submodule [TMP]
3. Implemented: Resource acquisition and loading for solid earth tides
   and ocean loading corrections. WIP implementation of the method that
   calculates the solid earth tide displacement [WIP]
4. Renamed `geophysical` submodule to `displacements`

Comments on 1
--------------

The idea is the same as for the delay models: we have a base class that
abstracts the concept of a geophysical displacement as something that
requires some resources to be present, some of them to be loaded and
then uses them to calculate a quantity. The idea is to be able to add
new corrections or delays without having to touch the main loop, which
just iterates over a set of objects of Delay/Displacement class.

Comments on 2
--------------

I really don't like the idea of the Constants class, the way it is
implemented or the way in which it identifies the version of the
ephemerides, but I am keeping it until the program works. I also assume
that there has to be a better way of doing these conversions within
spice.

Comments on 3
--------------

Although my initial intention (the one I had when I started working on
this branch) was to work with our own implementations of the models used
to calculate the geophysical displacements, I think that generating
extensions from the fortran versions disseminated with the IERS
Conventions using f2py is a better idea. I still like the concept of
having a base class for resource acquisition and all this, but I prefer
to use the `calculate` method to interface with the fortran routine and
leave the idea of a custom implementation for the future.

Comments on 4
---------------

None