EOPs and GCRS - ITRF conversion
- Exposed lagrange interpolator from INTERP.F to interpolate CIP offsets
  using the same algorithm as for the rest of the EOPs
- Implemented a class to provide access to interpolated (using INTERP.F)
  EOPs at a given epoch
- Implemented functions to get a rotation matrix between ITRF and GCRS
  (ERFA-based following IERS 2000/2006 algorithm)

NOTE 1: The algorithm for the conversion between ITRF and GCRS is the
same as in the original code, but I replaced the custom implementations
of the IAU routines with calls to ERFA functions. My implementation is
not validated yet.

NOTE 2: Astropy's EOPs (which are the ones I use here) are consistent
with ITRF 2020. I still need to verify this, but that should imply that
the transformation matrix mentioned above should go from ITRF2020 to
GCRS. The station coordinates in `glo.sit` are given in ITRF 2000, so we
should either update the coordinates to 2020 or use a routine to convert
between versions of ITRF.