Generate objects from experiment
Writing new code was starting to be difficult because the information
stored in the experiment object wasn't available for the objects it
contained (sources, stations...). In the new version, all the objects
are initialized from the experiment object directly and store a pointer
to the experiment as an attribute. This will make the remaining part of
the program easier to implement.